Syntax fixes, changed a tag to img tag
I wasn't sure if you were finished with this or not yet, but I looked through it to see if there was anything I could help with. It looks like you used an a tag instead of an img tag for the pictures, so they weren't showing up. There were also some missing closing arrows '>' in your html. 

The other thing I noticed was that you used different classes for each card. Ideally you would just use the same classes on each card since they are all basically the same style. That way you don't have to duplicate your code.